Moves NCBI formats rom kaiju to reference db
