+++ q2_types_genomics/reference_db/_format.py
@@ -8,6 +8,7 @@
 
 
 from qiime2.plugin import model
+from qiime2.core.exceptions import ValidationError
 from q2_types_genomics.plugin_setup import plugin
 from q2_types_genomics.reference_db._type import (
     ReferenceDB, Eggnog, Diamond, TaxonomyNCBI
@@ -58,24 +59,74 @@ plugin.register_semantic_type_to_format(ReferenceDB[Diamond],
                                         DiamondDatabaseDirFmt)
 
 
-class NCBITaxonomyTextFileFmt(model.TextFileFormat):
+class NCBITaxonomyNodesFormat(model.TextFileFormat):
+    def _validate_n_records(self, n=None):
+        with open(str(self), "r") as fh:
+            file_ = enumerate(fh) if n is None else zip(range(n), fh)
+
+            for i, line in file_:
+                line = line.rstrip("\n").split("\t|\t")
+                if 13 > len(line) or len(line) > 18:
+                    raise ValidationError(
+                        "NCBI taxonomy nodes file must have 13 columns, "
+                        f"found {len(line)} columns on line {i + 1}."
+                    )
+                if not line[0].isnumeric() or not line[1].isnumeric():
+                    raise ValidationError(
+                        "NCBI taxonomy nodes file must contain a numeric "
+                        "taxonomy ID in the first two columns, found "
+                        f"non-numeric value on line {i + 1}."
+                    )
+                for col in (5, 7, 9, 10, 11):
+                    if not line[col].isnumeric() or \
+                            not int(line[col]) in (0, 1):
+                        raise ValidationError(
+                            "NCBI taxonomy nodes file must contain 0 or 1 "
+                            "in columns 6, 8, 10, 11, and 12, found a "
+                            f"non-allowed value on line {i + 1}, column "
+                            f"{col + 1}: {line[col]}."
+                        )
+
     def _validate_(self, level):
-        # TODO: have native diamond validation run on db/self.path
-        pass
+        self._validate_n_records(n={"min": 10, "max": None}[level])
+
+
+class NCBITaxonomyNamesFormat(model.TextFileFormat):
+    def _validate_n_records(self, n=None):
+        with open(str(self), "r") as fh:
+            file_ = enumerate(fh) if n is None else zip(range(n), fh)
+
+            for i, line in file_:
+                line = line.rstrip("\n").split("\t|\t")
+                if len(line) != 4:
+                    raise ValidationError(
+                        "NCBI taxonomy names file must have 4 columns, "
+                        f"found {len(line)} columns on line {i + 1}."
+                    )
+                if not line[0].isnumeric():
+                    raise ValidationError(
+                        "NCBI taxonomy name file must contain a numeric "
+                        "taxonomy ID in the first column, found non-numeric "
+                        f"value on line {i + 1}: {line[0]}."
+                    )
+
+    def _validate_(self, level):
+        self._validate_n_records(n={"min": 10, "max": None}[level])
 
 
 class NCBITaxonomyBinaryFileFmt(model.TextFileFormat):
     def _validate_(self, level):
-        # TODO: have native diamond validation run on db/self.path
         pass
 
 
-plugin.register_formats(NCBITaxonomyTextFileFmt, NCBITaxonomyBinaryFileFmt)
+plugin.register_formats(
+    NCBITaxonomyNodesFormat, NCBITaxonomyNamesFormat, NCBITaxonomyBinaryFileFmt
+    )
 
 
 class NCBITaxonomyDirFmt(model.DirectoryFormat):
-    node = model.File('nodes.dmp', format=NCBITaxonomyTextFileFmt)
-    names = model.File('names.dmp', format=NCBITaxonomyTextFileFmt)
+    node = model.File('nodes.dmp', format=NCBITaxonomyNodesFormat)
+    names = model.File('names.dmp', format=NCBITaxonomyNamesFormat)
     tax_map = model.File(
         'prot.accession2taxid.gz',
         format=NCBITaxonomyBinaryFileFmt
